startup-report-generator/backend/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# PDF generation
weasyprint==62.3
markdown2==2.4.10
fonttools==4.58.5
cairocffi==1.7.1

# PDF text extraction
PyPDF2==3.0.1
pdfplumber==0.10.3

# AI and research
openai==1.3.7
requests==2.31.0

# Google integrations
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# Google Drive API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1

# Utility
python-dotenv==1.0.0
pydantic==2.5.0 